+++ ace-taskflow/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.30.0] - 2026-01-06
+
+### Added
+- Base36 compact ID support for idea directories
+- Base36 compact ID extraction from idea titles
+
 ## [0.29.1] - 2026-01-06
 
 ### Fixed